Rename library from gitlib4cpp to libgit4cpp
