+++ include/libgit4cpp/Error.h
@@ -1,8 +1,8 @@
 /**
- * \file   gitlib4cpp.h
+ * \file   Error.h
  * \author Lars Froehlich, Jannik Woehnert
  * \date   Created on June 7, 2023
- * \brief  Main include file for gitlib4cpp.
+ * \brief  Definition of the Error exception class.
  *
  * \copyright Copyright 2023 Deutsches Elektronen-Synchrotron (DESY), Hamburg
  *
@@ -22,12 +22,41 @@
 
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-#ifndef GITLIB4CPP_GITLIB4CPP_H_
-#define GITLIB4CPP_GITLIB4CPP_H_
+#ifndef LIBGIT4CPP_ERROR_H_
+#define LIBGIT4CPP_ERROR_H_
 
-#include "gitlib4cpp/Error.h"
+#include <stdexcept>
 
-/// Namespace git contains all functions, classes, and other declarations of gitlib4cpp.
-namespace git { }
+namespace git {
+
+/**
+ * An exception class carrying an error message.
+ *
+ * Error is used as the standard exception class throughout libgit4cpp.
+ * It can be used directly or inherited from.
+ *
+ * \code
+ * try
+ * {
+ *     throw git::Error("An error has occurred");
+ * }
+ * catch (const git::Error& e)
+ * {
+ *     std::cerr << e.what() << "\n";
+ * }
+ * \endcode
+ *
+ * \note
+ * git::Error is derived from std::runtime_error. It can therefore be caught by
+ * `catch (const std::exception&)`, `catch (const std::runtime_error&)`, and
+ * `catch (const git::Error&)`.
+ */
+class Error : public std::runtime_error
+{
+public:
+    using std::runtime_error::runtime_error;
+};
+
+} // namespace git
 
 #endif
