+++ include/libgit4cpp/GitRepository.h
