+++ include/libgit4cpp/LibGitPointer.h
